Move and generalize `ConfigProviderInterface`
